requirements.txt
Pillow~=9.2.0
requests~=2.28.1
cachetools~=5.2.0
python-dotenv~=0.20.0
numpy~=1.23.2
setuptools~=63.2.0
colorama~=0.4.5
opencv-python~=4.5.5.64
pynput~=1.7.6
pytesseract~=0.3.9